+++ javascript/UnixLepik.d.ts
@@ -196,6 +196,7 @@ declare class UnixLepik extends LepikEvents {
     setWindowPosition(windowId: string, x: number, y: number): void;
     focusNextWindow(): void;
     openApplication(command: string): void;
+    closeApplication(className: string): void;
     /**
      * Delays the execution for the specified number of milliseconds.
      * @param {number} ms - The number of milliseconds to delay.