+++ javascript/UnixLepik.js
@@ -306,6 +306,9 @@ class UnixLepik extends LepikEvents_1.LepikEvents {
     openApplication(command) {
         __classPrivateFieldGet(this, _UnixLepik_instances, "m", _UnixLepik_executeShellCommand).call(this, `xdotool exec ${command}`);
     }
+    closeApplication(className) {
+        __classPrivateFieldGet(this, _UnixLepik_instances, "m", _UnixLepik_executeShellCommand).call(this, `xdotool search --class "${className}" windowkill`);
+    }
     /**
      * Delays the execution for the specified number of milliseconds.
      * @param {number} ms - The number of milliseconds to delay.