+++ tests/singletons/openApplication.js
@@ -3,8 +3,6 @@ const lepik = require("../../index.js")
 async function main() {
   console.log(`Starting openApplication at ${Date.now()}`)
 
-  // Method implementation
-  console.log("opening application chrome")
   lepik.openApplication("chrome")
 
   lepik.close()